nhom4

+++ app/src/main/java/com/example/myapplication/DAO/HangDAO.java
@@ -5,6 +5,7 @@ import android.content.ContentValues;
 import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
+import android.widget.Toast;
 
 import com.example.myapplication.database.DBHelper;
 import com.example.myapplication.model.Hang;
@@ -35,8 +36,21 @@ public class HangDAO {
         String sql = "SELECT * FROM hang";
         return getData(sql);
     }
-    public long delete(String id) {
-        return db.delete("hang", "mahang = ?", new String[]{String.valueOf(id)});
+    public void delete(Context context, String id) {
+        ContentValues values = new ContentValues();
+        Cursor cursor = db.rawQuery("select * from sanpham where mahang = ?", new String[]{String.valueOf(id)});
+
+        if (cursor.getCount() != 0){
+            Toast.makeText(context, "Xóa không thành công do hãng có trong sản phẩm", Toast.LENGTH_SHORT).show();
+        }else{
+            long check = db.delete("hang", "mahang = ?", new String[]{String.valueOf(id)});
+
+            if (check > 0){
+                Toast.makeText(context, "Xóa thành công", Toast.LENGTH_SHORT).show();
+            }else{
+                Toast.makeText(context, "Xóa thất bại", Toast.LENGTH_SHORT).show();
+            }
+        }
     }
 
     public Hang getID(String id) {
